+++ src/components/input/InputColor.vue
@@ -1,27 +1,22 @@
 <template>
-	<Popover.Root>
-		<Popover.Trigger
-			class="inline-grid rounded grid-cols-2 grid-rows-1 w-full checkerboard h-full overflow-hidden"
-			:style="{ '--color-value': modelHex, '--alpha': modelAlpha }"
-		>
-			<div class="bg-(--color-value)"></div>
-			<div class="bg-(--color-value) opacity-(--alpha)"></div>
-		</Popover.Trigger>
-		<Popover.Portal>
-			<Popover.Content
-				avoid-collisions
-				position-strategy="absolute"
-				class="bg-ui-menu-bg-inner rounded p-2 z-1000"
+	<MxPopover>
+		<template #trigger>
+			<div
+				class="inline-grid rounded grid-cols-2 grid-rows-1 w-full checkerboard h-full
+					overflow-hidden"
+				:style="{ '--color-value': modelHex, '--alpha': modelAlpha }"
 			>
-				<ColorPicker v-model="modelHex" v-model:alpha="modelAlpha" :transparency />
-			</Popover.Content>
-		</Popover.Portal>
-	</Popover.Root>
+				<div class="bg-(--color-value)"></div>
+				<div class="bg-(--color-value) opacity-(--alpha)"></div>
+			</div>
+		</template>
+		<template #content>
+			<ColorPicker v-model="modelHex" v-model:alpha="modelAlpha" :transparency />
+		</template>
+	</MxPopover>
 </template>
 
 <script lang="ts" setup>
-import { Popover } from 'reka-ui/namespaced'
-
 defineProps<{
 	transparency?: boolean
 }>()